+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/ImmutableConfig.kt
@@ -5,6 +5,7 @@ import android.content.Context
 import android.os.Build
 import com.bugsnag.android.performance.AutoInstrument
 import com.bugsnag.android.performance.Logger
+import com.bugsnag.android.performance.NetworkRequestInstrumentationCallback
 import com.bugsnag.android.performance.PerformanceConfiguration
 
 internal data class ImmutableConfig(
@@ -19,6 +20,7 @@ internal data class ImmutableConfig(
     val versionCode: Long?,
     val appVersion: String?,
     val logger: Logger,
+    var networkRequestCallback: NetworkRequestInstrumentationCallback?,
 ) {
     val isReleaseStageEnabled =
         enabledReleaseStages == null || enabledReleaseStages.contains(releaseStage)
@@ -37,6 +39,7 @@ internal data class ImmutableConfig(
         configuration.logger
             ?: if (getReleaseStage(configuration) == RELEASE_STAGE_PRODUCTION) NoopLogger
             else DebugLogger,
+        configuration.networkRequestCallback,
     )
 
     companion object {