+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/InstrumentedAppState.kt
@@ -54,6 +54,7 @@ class InstrumentedAppState {
 
         spanProcessor = tracer
         spanFactory.spanProcessor = tracer
+        spanFactory.networkRequestCallback = configuration.networkRequestCallback
 
         if (configuration.autoInstrumentAppStarts) {
             // redirect existing spanProcessor -> new Tracer